called weather api from blitz-build-be
