--- blitz-build/.env
@@ -1,1 +1,0 @@
-REACT_APP_DARKSKY_WEATHER_SECRET_KEY = "425ccdff13ee9871b8d3b2ad4fb9632f"
