+++ blitz-build/src/components/auth/Login.jsx
@@ -1,13 +1,13 @@
 import React from "react";
-import { Link } from "react-router-dom"
-
-import DashboardWeater from '../weather/DashboardWeather.jsx'
+import { Link } from "react-router-dom";
 
 function Login() {
   return (
     <div>
       <h1>Login</h1>
-      <p>Already have an account? <Link to="/signup"> Sign up</Link></p>
+      <p>
+        Already have an account? <Link to="/signup"> Sign up</Link>
+      </p>
       <form>
         <label>Email</label>
         <input placeholder="john.doe@gmail.com" />
@@ -15,10 +15,8 @@ function Login() {
         <input placeholder="••••••••" />
         <button>Log in</button>
       </form>
-      <DashboardWeater />
     </div>
   );
 }
 
 export default Login;
-
